first test made, adjusting sbt to run junit tests
