+++ build.sbt
@@ -14,6 +14,13 @@ javacOptions ++= Seq("-source", "8", "-target", "8")
 
 javacOptions in doc := Seq("-source", "8")
 
+//projectDependencies ++= Seq(
+//"org.scalatest" %% "scalatest" % "2.2.2" % Test,
+//"junit" % "junit" % "4.11" % Test,
+//crossPaths := false,
+//"com.novocode" % "junit-interface" % "0.11" % Test
+//)
+
 libraryDependencies ++= Seq(
   "com.google.code.gson" % "gson" % "2.6.2",
   "org.apache.httpcomponents" % "httpasyncclient" % "4.1.1",
@@ -24,3 +31,5 @@ libraryDependencies ++= Seq(
 
 // Disable using the Scala version in published artifacts
 crossPaths := false
+
+testOptions in Test := Seq(Tests.Argument(TestFrameworks.JUnit, "-a"))
