+++ src/test/java/AlgorithmHandlerTest.java
@@ -1,59 +1,32 @@
-import com.algorithmia.Algorithmia;
-import com.algorithmia.AlgorithmiaClient;
+
 import com.algorithmia.algorithmHandler.*;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParser;
-
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.lang.Integer;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Arrays;
-
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.Assert;
 
-
-
-
-class AdvancedInput{
-    String name;
-    Integer age;
-}
-
-class BasicAlgroithm{
-    String Foo(String input) throws Exception {
-        return "Hello " + input;
-    }
-}
-
-class AdvancedAlgorithm {
-    String Apply(AdvancedInput input, HashMap<String, String> context) throws Exception{
-        if(context.containsKey("local_file")){
-            return "Hello " + input.name + " you are " + input.age +
-                    " years old, and your model file is downloaded here " + context.get("local_file");
-        }
-        return "hello " + input.name+ " you are " + input.age + " years old";
-    }
-    HashMap<String, String> DownloadModel() throws Exception{
-        HashMap<String, String> context = new HashMap<>();
-        context.put("local_file", "/tmp/somefile");
-        return context;
-    }
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
 
 
 public class AlgorithmHandlerTest {
+
+    @Before
+    public void IntializePipe(){
+        
+    }
+
     /// TEXT hello world
     @Test
-    public void TestSimple(){
-        BasicAlgroithm algo = new BasicAlgroithm();
+    public void SimpleAlgoRequest(){
+        BasicAlgorithm algo = new BasicAlgorithm();
         AlgorithmHandler handler = new AlgorithmHandler<>(algo::Foo);
         try {
+            InputStream fakeIn = new ByteArrayInputStream("{\"content_type\":\"text\", \"data\":\"james\")\n".getBytes());
+            System.setIn(fakeIn);
             handler.run();
+
         } catch(Exception e){
             Assert.fail();
         }
     }
-}
 }
